src/requirements.txt
asgiref==3.8.1
astroid==3.3.9
attrs==25.1.0
autopep8==2.3.2
dill==0.3.9
Django==5.1.7
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-tailwind==3.8.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mysqlclient==2.2.7
packaging==24.2
platformdirs==4.3.6
pycodestyle==2.12.1
pylint==3.3.5
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tomlkit==0.13.2
uritemplate==4.1.1
